Fix clang-tidy violations in category/execution
Fix all directly-fixable clang-tidy violations found when expanding
lint coverage to category/execution/. This enables 20 of the 26
originally-suppressed checks. Remaining 2 temporary suppressions
(misc-auto-const-correctness, misc-include-cleaner) will be addressed
in follow-up commits.

Changes by check:

  modernize-use-nullptr: NULL → nullptr in secp256k1 call
  performance-no-automatic-move: drop const from receipt_bytes to
    allow implicit move on return
  modernize-use-transparent-functors: std::greater<uint64_t> →
    std::greater<>
  bugprone-too-small-loop-variable: uint8_t → size_t for loop
    bounded by sizeof
  bugprone-string-constructor: NOLINT — intentional empty
    string_view to avoid null-pointer UB in silkpre
  bugprone-switch-missing-default-case: add default: to syscall
    signature switch
  bugprone-exception-escape: wrap TraceTimer dtor in try/catch;
    NOLINT on fuzzer main()
  modernize-loop-convert: iterator loops → range-based for with
    structured bindings in DbCache
  performance-noexcept-move-constructor: add noexcept to BlockDb
    and FileDb move constructors
  performance-move-const-arg: remove std::move on const refs
    (logs, hash); make lambda params non-const to enable move
  performance-unnecessary-copy-initialization: const → const& for
    copies that are never modified
  performance-inefficient-vector-operation: add reserve() before
    loops
  modernize-use-emplace: push_back → emplace_back where possible;
    NOLINT where aggregate types prevent in-place construction
  google-explicit-constructor: add explicit to 6 single-argument
    constructors (DbCache, StakingContractMachine, BLS/secp types)
  llvm-include-order: reorder gtest includes per LLVM priority
    categories (gtest/gmock before other third-party headers)
  readability-qualified-auto: auto → auto *const for pointer types
    (raw string literals, json strings)
  bugprone-use-after-move: NOLINT on intentional vector reuse
    after move in db/util.cpp; fix multiple-forward in
    vm/fuzzing/generator/choice.hpp
  bugprone-suspicious-memory-comparison: NOLINT on intentional
    binary memcmp in event recorder tests

Permanently suppressed checks (false-positive-heavy on this codebase):
  - bugprone-suspicious-stringview-data-usage: binary data
    processing uses .data() with explicit size everywhere
  - bugprone-macro-parentheses: macro args used as template/member
    names cannot be parenthesized
  - clang-analyzer-core.CallAndMessage: can't trace through C
    library calls (BLS/secp256k1)
  - clang-analyzer-core.uninitialized.Assign: false positives on
    default-initialized structs and boost system headers
  - clang-analyzer-cplusplus.NewDeleteLeaks: false positive on
    complex control flow in db traversal
  - clang-analyzer-unix.MallocSizeof: fixed by using named
    constant for allocation size

+++ category/execution/ethereum/core/contract/test_big_endian.cpp
@@ -41,7 +41,7 @@ TYPED_TEST(BigEndianEncodeTest, uint_max)
     constexpr NativeType native = [] {
         NativeType v{};
         // generate a sequence that has a different big endian representation
-        for (uint8_t i = 1; i <= sizeof(NativeType); ++i) {
+        for (size_t i = 1; i <= sizeof(NativeType); ++i) {
             v = static_cast<NativeType>((v << 8) | i);
         }
         return v;
